Store descendant depth on `#rebuild!`
Add a depth column to the hierarchy table to track
the depth of a path on all descendant entries for
that path. For instance, if we have

  a => b => c

We store one row for a, two rows for b, and three 
rows for c. Now we have a depth count for 
each of these rows. 

  a depth is 1, b is 2, and c is 3.

This would look something like this:

item  parent  anc  desc gens depth
========================================
  a    null    a    a     0    1
  b     a      b    b     0    2
  b     a      a    b     1    2
  c     b      c    c     0    3
  c     b      b    c     1    3
  c     b      a    c     2    3

Having depth makes it trivially easy to limit 
descendant depth in queries. At any point in the
path we know the total depth.