Update `#find_all_by_generation`
Replace the INNER JOIN with a simple subselect. 
Since we know that the current model id is the
ancestor we don't need a complicated query to 
scan the descendants. Again, this is probably a 
modest speed improvement that will become more
apparent with a large amount of data (we now don't
have to worry about the joined rows). What we 
gain is a method that is easier to work with and
easier to chain with other scopes.